etl/requirements.txt
pdfplumber>=0.11.0
python-Levenshtein>=0.25.0
